+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 2.4.1 (unreleased)
 ------------------
 
--   Add GWSkyNet 2.4.1 to avoid installing tensorflow-cpu >= 2.16.1. 
+-   Add GWSkyNet 2.4.1 to avoid installing tensorflow-cpu >= 2.16.1.
     Clairfy the version of annotated skymaps and quantities in the uploads.
     Add public tag to GWSkyNet's uploads.
 
@@ -31,6 +31,8 @@ Changelog
     update notices
 
 -   Serialize changing of preferred external event.
+
+-   Update Nagios URL and screenshot in the documentation.
 
 2.3.6 "Champ" (2024-04-30)
 --------------------------